requirements.txt
django>=5.0,<6
fastapi==0.109.2
uvicorn==0.27.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator==2.1.1
python-multipart==0.0.9
python-dotenv==1.0.1
dj-database-url==2.1.0
psycopg[binary]==3.3.3
